Add files via upload
This script uses the NTFSSecurity module and adds a menu with four options for managing local users and groups. The script prompts the user for input, such as creating a new local user, adding or removing users from groups, or removing an existing local user. The switch statement performs the selected action based on the input provided by the user.

Note that as you might know, depending on the environment, you must make changes to the script in order for it to work properly.